DPC-4146 Reskin AO Happy Path (#2219)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4146

## 🛠 Changes

- Step indicator component moved to h1
- Invitation login component moved to invitations module
- Invitation components skinned to match sketch and content
- User signed in after completing registration

## ℹ️ Context

We got the
[comps](https://www.sketch.com/s/fa3f816d-9a9a-4808-a7fe-e509a74e125c/p/615FBE78-4752-4010-B83F-BC512A1D1CE2/canvas)
and [content](https://confluence.cms.gov/x/mJ_EOw) from design.

## 🧪 Validation

Manual and automated testing

### Screenshots
![Screenshot 2024-07-22 at 5 22
16 PM](https://github.com/user-attachments/assets/4d28435f-f883-4376-81a4-9a887eb2d25a)
![Screenshot 2024-07-25 at 3 52
35 PM](https://github.com/user-attachments/assets/a431c3ba-4a44-479a-8ef5-df1ccce18a5c)
![Screenshot 2024-07-22 at 5 23
02 PM](https://github.com/user-attachments/assets/f4d580ae-dd2a-4162-b9ac-8a30180a0fa3)
![Screenshot 2024-07-22 at 5 23
24 PM](https://github.com/user-attachments/assets/f0dd0c20-4762-42c0-a387-eefcb10efdd7)
![Screenshot 2024-07-22 at 5 23
44 PM](https://github.com/user-attachments/assets/e0e40726-1084-4059-bf3d-ea5df2d3b6de)

+++ dpc-portal/app/controllers/invitations_controller.rb
@@ -15,7 +15,7 @@ class InvitationsController < ApplicationController
 
   def accept
     session["invitation_status_#{@invitation.id}"] = 'identity_verified'
-    render(Page::Invitations::AcceptInvitationComponent.new(@organization, @invitation))
+    render(Page::Invitations::AcceptInvitationComponent.new(@organization, @invitation, @given_name, @family_name))
   end
 
   def confirm
@@ -29,15 +29,10 @@ class InvitationsController < ApplicationController
                                                             @invitation)
     end
 
-    if @invitation.credential_delegate?
-      create_cd_org_link
-    elsif @invitation.authorized_official?
-      create_ao_org_link
-    else
-      return render(Page::Invitations::BadInvitationComponent.new(@invitation, 'invalid', 'warning'),
-                    status: :unprocessable_entity)
-    end
+    return unless create_link
+
     session.delete("invitation_status_#{@invitation.id}")
+    sign_in(:user, @user)
     render(Page::Invitations::SuccessComponent.new(@organization, @invitation))
   end
 
@@ -67,26 +62,36 @@ class InvitationsController < ApplicationController
 
   private
 
+  def create_link
+    if @invitation.credential_delegate?
+      create_cd_org_link
+    elsif @invitation.authorized_official?
+      create_ao_org_link
+    else
+      render(Page::Invitations::BadInvitationComponent.new(@invitation, 'invalid', 'warning'),
+             status: :unprocessable_entity)
+      false
+    end
+  end
+
   def create_cd_org_link
     CdOrgLink.create!(user:, provider_organization: @organization, invitation: @invitation)
     @invitation.accept!
-    flash[:notice] = "Invitation accepted. You can now manage this organization's credentials. Learn more."
   end
 
   def create_ao_org_link
     AoOrgLink.create!(user:, provider_organization: @organization, invitation: @invitation)
     @invitation.accept!
-    flash[:notice] = 'Invitation accepted.'
   end
 
   def user
     user_info = UserInfoService.new.user_info(session)
-    local_user = User.find_or_create_by!(provider: :openid_connect, uid: user_info['sub']) do |user_to_create|
+    @user = User.find_or_create_by!(provider: :openid_connect, uid: user_info['sub']) do |user_to_create|
       user_to_create.email = @invitation.invited_email
       user_to_create.pac_id = session.delete(:user_pac_id)
     end
-    local_user.update(pac_id: session.delete(:user_pac_id)) unless local_user.pac_id
-    local_user
+    @user.update(pac_id: session.delete(:user_pac_id)) unless @user.pac_id
+    @user
   end
 
   def check_for_token
@@ -96,7 +101,7 @@ class InvitationsController < ApplicationController
       return
     end
 
-    render(Page::Session::InvitationLoginComponent.new(@invitation))
+    render(Page::Invitations::InvitationLoginComponent.new(@invitation))
   end
 
   def invitation_matches_user
@@ -105,6 +110,8 @@ class InvitationsController < ApplicationController
       render(Page::Invitations::BadInvitationComponent.new(@invitation, 'pii_mismatch', 'error'),
              status: :forbidden)
     end
+    @given_name = user_info['given_name']
+    @family_name = user_info['family_name']
   rescue UserInfoServiceError => e
     handle_user_info_service_error(e)
   end
@@ -127,7 +134,7 @@ class InvitationsController < ApplicationController
     return if params[:verification_code] == @invitation.verification_code
 
     @invitation.errors.add(:verification_code, :bad_code, message: 'tbd')
-    render(Page::Invitations::AcceptInvitationComponent.new(@organization, @invitation),
+    render(Page::Invitations::AcceptInvitationComponent.new(@organization, @invitation, @given_name, @family_name),
            status: :bad_request)
   end
 
@@ -141,7 +148,7 @@ class InvitationsController < ApplicationController
   def handle_user_info_service_error(error)
     case error.message
     when 'unauthorized'
-      render(Page::Session::InvitationLoginComponent.new(@invitation))
+      render(Page::Invitations::InvitationLoginComponent.new(@invitation))
     else
       render(Page::Invitations::BadInvitationComponent.new(@invitation, 'server_error', 'warning'),
              status: :service_unavailable)
